+++ plugins/mobile-apps/template/babel.config.js
@@ -1,3 +1,5 @@
+const { createPowerAppsBabelConfig } = require('@microsoft/power-apps-native-host/config/babelConfig');
+
 // CUSTOMIZATION START - DO NOT REMOVE OR RENAME THE COMMENT
 // Add Babel presets and plugins to these arrays only.
 const customPresets = [];
@@ -5,13 +7,8 @@ const customPlugins = [];
 // CUSTOMIZATION END - DO NOT REMOVE OR RENAME THE COMMENT
 
 module.exports = function (api) {
-  api.cache(true);
-  return {
-    presets: ['babel-preset-expo', ...customPresets],
-    plugins: [
-      ...customPlugins,
-      // Must be last
-      'react-native-reanimated/plugin',
-    ],
-  };
+  return createPowerAppsBabelConfig(api, {
+    presets: customPresets,
+    plugins: customPlugins,
+  });
 };